FastAPI changes
