+++ {{cookiecutter.__src_folder_name}}/.dockerignore
@@ -1,0 +1,3 @@
+.git*
+.venv/
+**/*.pyc
