+++ {{cookiecutter.__src_folder_name}}/Dockerfile
@@ -16,7 +16,7 @@ COPY entrypoint.sh .
 RUN chmod +x entrypoint.sh
 
 COPY . .
-{% if cookiecutter.project_backend == "flask" %}
+{% if cookiecutter.project_backend in ("fastapi", "flask") %}
 RUN python -m pip install -e .
 {% endif %}
 CMD bash -c ". entrypoint.sh"
