+++ {{cookiecutter.__src_folder_name}}/fastapi/fastapi_app/__init__.py
@@ -1,0 +1,1 @@
+from .app import app
