+++ {{cookiecutter.__src_folder_name}}/requirements.txt
@@ -6,15 +6,6 @@ gunicorn==20.1.0
 psycopg2-binary==2.9.6
 {% endif %}
 
-{% if cookiecutter.project_backend == "fastapi" %}
-# fastapi requirements
-fastapi
-sqlmodel
-jinja2
-uvicorn[standard]
-python-multipart
-{% endif %}
-
 {% if cookiecutter.project_backend == "django" %}
 # django requirements
 django
