+++ {{cookiecutter.__src_folder_name}}/root/README.md
@@ -1,3 +1,10 @@
+{% if cookiecutter.project_backend == "fastapi" %}
+# Deploy a FastAPI Application via Azure Container Apps
+
+This project deploy a [FastAPI](https://fastapi.tiangolo.com) application to [Azure Container Apps](https://aka.ms/aca). The FastAPI application is a simple web application for a space travel agency. The application is built using the FastAPI framework and uses a PostgreSQL database with SQLModel as an ORM. The application can be deployed to Azure using the [Azure Developer CLI](https://learn.microsoft.com/azure/developer/azure-developer-cli/overview).
+
+{% endif %}
+
 ## Opening the project
 
 This project has [Dev Container support](https://code.visualstudio.com/docs/devcontainers/containers), so it will be be setup automatically if you open it in Github Codespaces or in local VS Code with the [Dev Containers extension](https://marketplace.visualstudio.com/items?itemName=ms-vscode-remote.remote-containers).
@@ -12,35 +19,31 @@ If you're not using one of those options for opening the project, then you'll ne
     python -m pip install -r {{cookiecutter.__src_folder_name}}/requirements.txt
     ```
 
-{% if cookiecutter.project_backend == "django" %}
+{% if cookiecutter.project_backend in ("flask", "fastapi") %}
+1. Install the app as an editable package:
+
+    ```sh
+    python -m pip install -e {{cookiecutter.__src_folder_name}}
+    ```
+{% endif %}
+
 1. Apply database migrations and seed initial data:
 
+{% if cookiecutter.project_backend == "django" %}
     ```sh
     python {{cookiecutter.__src_folder_name}}/manage.py migrate
     python {{cookiecutter.__src_folder_name}}/manage.py loaddata {{cookiecutter.__src_folder_name}}/seed_data.json
     ```
 {% endif %}
-
 {% if cookiecutter.project_backend == "flask" %}
-1. Install the Flask app as an editable package:
-
-    ```sh
-    python -m pip install -e {{cookiecutter.__src_folder_name}}
-    ```
-
-1. Apply database migrations and seed initial data:
-
     ```sh
     python3 -m flask --app flaskapp db upgrade --directory {{cookiecutter.__src_folder_name}}/flaskapp/migrations
     python3 -m flask --app flaskapp seed --filename {{cookiecutter.__src_folder_name}}/seed_data.json
     ```
 {% endif %}
-
 {% if cookiecutter.project_backend == "fastapi" %}
-1. Apply database migrations and seed initial data:
-    
     ```sh
-    python3 seed_data.py
+    python3 demo_code/fastapi_app/seed_data.py
     ```
 {% endif %}
 
@@ -50,10 +53,10 @@ Run gunicorn on the app:
 
 ```sh
 {% if cookiecutter.project_backend == "flask" %}
-python3 -m gunicorn 'flaskapp:create_app()'
+python3 -m gunicorn 'flaskapp:create_app()' -c demo_code/gunicorn.conf.py
 {% endif %}
 {% if cookiecutter.project_backend == "fastapi" %}
-python3 -m gunicorn app:app
+python3 -m gunicorn fastapi_app:app -c demo_code/gunicorn.conf.py
 {% endif %}
 {% if cookiecutter.project_backend == "django" %}
 python3 manage.py collectstatic
